+++ src/MixedModels.jl
@@ -16,6 +16,7 @@ using RectangularFullPacked
 using SparseArrays
 using StaticArrays
 using Statistics
+using StatsAPI
 using StatsBase
 using StatsModels
 using StructTypes
@@ -23,16 +24,11 @@ using Tables
 
 using LinearAlgebra: BlasFloat, BlasReal, HermOrSym, PosDefException, copytri!
 using Base: Ryu, require_one_based_indexing
-using GLM: Link, canonicallink, linkfun, linkinv
+using DataAPI: levels, refpool, refarray, refvalue
+using GLM: Link, canonicallink, linkfun, linkinv, dispersion, dispersion_parameter
+using NLopt: Opt
 using StatsModels: TableRegressionModel
-
 using StatsFuns: log2π, normccdf
-
-import Base: *
-import DataAPI: levels, refpool, refarray, refvalue
-import GLM: dispersion, dispersion_parameter
-import NLopt: Opt
-import StatsBase: fit, fit!
 
 export @formula,
     AbstractReMat,
@@ -131,13 +127,12 @@ export @formula,
     sparseL,
     std,
     stderror,
+    stderror!,
     updateL!,
     varest,
     vcov,
     weights,
     zerocorr
-
-import Base: ==, *
 
 """
     MixedModel