+++ src/linearmixedmodel.jl
@@ -46,6 +46,9 @@ function LinearMixedModel(
 )
     return LinearMixedModel(f::FormulaTerm, Tables.columntable(tbl); contrasts, wts, σ)
 end
+const _MISSING_RE_ERROR = ArgumentError(
+    "Formula contains no random effects; this isn't a mixed model. Perhaps you want to use GLM.jl?",
+)
 
 function LinearMixedModel(
     f::FormulaTerm, tbl::Tables.ColumnTable; contrasts=Dict{Symbol,Any}(), wts=[], σ=nothing
@@ -62,6 +65,11 @@ function LinearMixedModel(
         rethrow(e)
     end
     form = apply_schema(f, sch, LinearMixedModel)
+
+    if form.rhs isa MatrixTerm || !any(x -> isa(x, AbstractReTerm), form.rhs)
+        throw(_MISSING_RE_ERROR)
+    end
+
     # tbl, _ = StatsModels.missing_omit(tbl, form)
 
     y, Xs = modelcols(form, tbl)
@@ -119,6 +127,7 @@ function LinearMixedModel(
             push!(feterms, FeTerm(x, isa(cnames, String) ? [cnames] : collect(cnames)))
         end
     end
+    isempty(reterms) && throw(_MISSING_RE_ERROR)
     return LinearMixedModel(convert(Array{T}, y), only(feterms), reterms, form, wts, σ)
 end
 
@@ -174,7 +183,7 @@ function LinearMixedModel(
     )
 end
 
-function fit(
+function StatsAPI.fit(
     ::Type{LinearMixedModel},
     f::FormulaTerm,
     tbl;
@@ -198,7 +207,7 @@ function fit(
     )
 end
 
-function fit(
+function StatsAPI.fit(
     ::Type{LinearMixedModel},
     f::FormulaTerm,
     tbl::Tables.ColumnTable;
@@ -220,7 +229,7 @@ function _offseterr()
     )
 end
 
-function fit(
+function StatsAPI.fit(
     ::Type{MixedModel},
     f::FormulaTerm,
     tbl;
@@ -239,7 +248,7 @@ function fit(
     end
 end
 
-function fit(
+function StatsAPI.fit(
     ::Type{MixedModel},
     f::FormulaTerm,
     tbl,
@@ -265,19 +274,19 @@ function fit(
     end
 end
 
-function StatsBase.coef(m::LinearMixedModel{T}) where {T}
+function StatsAPI.coef(m::LinearMixedModel{T}) where {T}
     piv = m.feterm.piv
     return invpermute!(fixef!(similar(piv, T), m), piv)
 end
 
 βs(m::LinearMixedModel) = NamedTuple{(Symbol.(coefnames(m))...,)}(coef(m))
 
-function StatsBase.coefnames(m::LinearMixedModel)
+function StatsAPI.coefnames(m::LinearMixedModel)
     Xtrm = m.feterm
     return invpermute!(copy(Xtrm.cnames), Xtrm.piv)
 end
 
-function StatsBase.coeftable(m::LinearMixedModel)
+function StatsAPI.coeftable(m::LinearMixedModel)
     co = coef(m)
     se = stderror!(similar(co), m)
     z = co ./ se
@@ -387,7 +396,7 @@ function createAL(reterms::Vector{AbstractReMat{T}}, Xy::FeMat{T}) where {T}
     return A, L
 end
 
-StatsBase.deviance(m::LinearMixedModel) = objective(m)
+StatsAPI.deviance(m::LinearMixedModel) = objective(m)
 
 GLM.dispersion(m::LinearMixedModel, sqr::Bool=false) = sqr ? varest(m) : sdest(m)
 
@@ -418,7 +427,7 @@ objective, if the optimization takes more than one second or so.
 At every `thin`th iteration  is recorded in `fitlog`, optimization progress is
 saved in `m.optsum.fitlog`.
 """
-function fit!(
+function StatsAPI.fit!(
     m::LinearMixedModel{T};
     progress::Bool=true,
     REML::Bool=false,
@@ -512,6 +521,13 @@ function fit!(
     return m
 end
 
+"""
+    fitted!(v::AbstractArray{T}, m::LinearMixedModel{T})
+
+Overwrite `v` with the fitted values from `m`.
+
+See also `fitted`.
+"""
 function fitted!(v::AbstractArray{T}, m::LinearMixedModel{T}) where {T}
     ## FIXME: Create and use `effects(m) -> β, b` w/o calculating β twice
     Xtrm = m.feterm
@@ -522,7 +538,7 @@ function fitted!(v::AbstractArray{T}, m::LinearMixedModel{T}) where {T}
     return v
 end
 
-StatsBase.fitted(m::LinearMixedModel{T}) where {T} = fitted!(Vector{T}(undef, nobs(m)), m)
+StatsAPI.fitted(m::LinearMixedModel{T}) where {T} = fitted!(Vector{T}(undef, nobs(m)), m)
 
 """
     fixef!(v::Vector{T}, m::MixedModel{T})
@@ -650,7 +666,7 @@ function Base.getproperty(m::LinearMixedModel{T}, s::Symbol) where {T}
     end
 end
 
-StatsBase.islinear(m::LinearMixedModel) = true
+StatsAPI.islinear(m::LinearMixedModel) = true
 
 """
     _3blockL(::LinearMixedModel)
@@ -701,7 +717,7 @@ for the model in the sense that this sum is the dimension of the span of columns
 of the model matrix.  With a bit of hand waving a similar argument could be made
 for linear mixed-effects models. The hat matrix is of the form ``[ZΛ X][L L']⁻¹[ZΛ X]'``.
 """
-function StatsBase.leverage(m::LinearMixedModel{T}) where {T}
+function StatsAPI.leverage(m::LinearMixedModel{T}) where {T}
     # To obtain the diagonal elements solve L⁻¹[ZΛ X]'eⱼ
     # where eⱼ is the j'th basis vector in Rⁿ and evaluate the squared length of the solution.
     # The fact that the [1,1] block of L is always UniformBlockDiagonal
@@ -742,7 +758,7 @@ function StatsBase.leverage(m::LinearMixedModel{T}) where {T}
     return value
 end
 
-function StatsBase.loglikelihood(m::LinearMixedModel)
+function StatsAPI.loglikelihood(m::LinearMixedModel)
     if m.optsum.REML
         throw(ArgumentError("loglikelihood not available for models fit by REML"))
     end
@@ -1151,7 +1167,7 @@ function stderror!(v::AbstractVector{Tv}, m::LinearMixedModel{T}) where {Tv,T}
     return v
 end
 
-function StatsBase.stderror(m::LinearMixedModel{T}) where {T}
+function StatsAPI.stderror(m::LinearMixedModel{T}) where {T}
     return stderror!(similar(m.feterm.piv, T), m)
 end
 
@@ -1205,14 +1221,14 @@ This is the crucial step in evaluating the objective, given a new parameter valu
 function updateL!(m::LinearMixedModel{T}) where {T}
     A, L, reterms = m.A, m.L, m.reterms
     k = length(reterms)
-    for (l, a) in zip(L, A) # copy each element of A to corresponding element of L
-        copyto!(l, a)       # For some reason copyto!.(L, A) allocates a lot of memory
-    end
+    copyto!(last(m.L), last(m.A))  # ensure the fixed-effects:response block is copied
     for j in eachindex(reterms) # pre- and post-multiply by Λ, add I to diagonal
         cj = reterms[j]
-        scaleinflate!(L[kp1choose2(j)], cj)
+        diagind = kp1choose2(j)
+        copyscaleinflate!(L[diagind], A[diagind], cj)
         for i in (j + 1):(k + 1)     # postmultiply column by Λ
-            rmulΛ!(L[block(i, j)], cj)
+            bij = block(i, j)
+            rmulΛ!(copyto!(L[bij], A[bij]), cj)
         end
         for jj in 1:(j - 1)        # premultiply row by Λ'
             lmulΛ!(cj', L[block(j, jj)])
@@ -1245,7 +1261,7 @@ function varest(m::LinearMixedModel)
     return isnothing(m.optsum.sigma) ? pwrss(m) / ssqdenom(m) : m.optsum.sigma
 end
 
-function StatsBase.weights(m::LinearMixedModel)
+function StatsAPI.weights(m::LinearMixedModel)
     rtwts = m.sqrtwts
     return isempty(rtwts) ? ones(eltype(rtwts), nobs(m)) : abs2.(rtwts)
 end